momoize all render component and fix missing calendar
